# openapi.yaml
openapi: 3.0.3
info:
  title: Efficient AI API
  description: |
    OpenAI-compatible API with x402 micropayments. 
    Routes 88% of queries to embedded engine for $0.0001.
    No API key needed — pay per request via x402.
  version: 0.1.0
  contact:
    name: Efficient AI
    url: https://github.com/efficient-ai/efficient-ai
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.efficient-ai.com
    description: Production
paths:
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      operationId: createChatCompletion
      description: |
        Creates a model response for the given chat conversation.
        Supports x402 payment protocol for monetization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                model:
                  type: string
                  default: auto
                  description: Model to use. 'auto' for automatic routing.
                messages:
                  type: array
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
                temperature:
                  type: number
                  minimum: 0
                  maximum: 2
                  default: 0.7
                max_tokens:
                  type: integer
                  minimum: 1
                stream:
                  type: boolean
                  default: false
                response_format:
                  type: object
                  description: Format for structured output
      responses:
        '200':
          description: Successful response
          headers:
            X-Tier:
              description: Backend used (engine/ollama/cloud)
              schema:
                type: string
            X-Price:
              description: Price charged for this request
              schema:
                type: string
            X-Model:
              description: Model used
              schema:
                type: string
            X-Latency-ms:
              description: Request latency in milliseconds
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  object:
                    type: string
                  created:
                    type: integer
                  model:
                    type: string
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        message:
                          type: object
                          properties:
                            role:
                              type: string
                            content:
                              type: string
                        finish_reason:
                          type: string
                          enum: [stop, length, content_filter]
                  usage:
                    type: object
                    properties:
                      prompt_tokens:
                        type: integer
                      completion_tokens:
                        type: integer
                      total_tokens:
                        type: integer
        '402':
          description: Payment Required (x402)
          headers:
            PAYMENT-REQUIRED:
              description: Base64-encoded payment requirements
              schema:
                type: string
            X-Price:
              description: Price in USD
              schema:
                type: string
            X-Tier:
              description: Backend tier
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  scheme:
                    type: string
                  network:
                    type: string
                  pay_to:
                    type: string
                  price:
                    type: object
                    properties:
                      amount:
                        type: string
                      asset:
                        type: string
                      extra:
                        type: object
                  description:
                    type: string
                  mimeType:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                      type:
                        type: string
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                      type:
                        type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                      type:
                        type: string
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    x402Payment:
      type: apiKey
      in: header
      name: PAYMENT-SIGNATURE
      description: x402 payment signature for monetization
security:
  - ApiKeyAuth: []
  - x402Payment: []